+++ CHANGELOG.md
@@ -37,6 +37,7 @@ This update introduces some new tables. Refer to [UPGRADE.md](UPGRADE.md) for th
 4. Moved static text to translation file. (issue #102)
 5. Add getTotaldiscount method to Order. (issue #142)
 6. Improved Tinymce editor to match RedminPortal look and feel.
+7. Updated Tinymce to version 4.3.2.
 
 ### Bug fixes:
 1. User deletion: prevent user from deleting or deactivating own account while they are logged in. (issue #136)